Add color highlighting support for YAML files
Extract the color map from r_highlight_rules.js into a shared colors.js
module and reuse it in yaml_highlight_rules.js to provide color preview
for hex colors (#rrggbb, #rrggbbaa, #rgb) and named R colors in YAML
files like _pkgdown.yml and _brand.yml.

Supports both quoted strings ("red", "#ff0000") and unquoted named
colors (red, steelblue4) in YAML files. Unquoted hex colors are not
supported since '#' starts a comment in YAML.

Fixes #16657
